Move ValueClassBoxConverter to use serialization and deserialization
